+++ CHANGELOG.md
@@ -10,6 +10,7 @@ that users understand how the changes affect the new version.
 
 v0.0.1
 ---------------------------
++ Add the number of alleles per marker to the final report.
 + Add allele frequencies to the final report.
 + Add support for multiple fastq files per sample via PEP subsample
 configuration.